backend/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
Django==5.1.2
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.15.4
